first pass at trying to fix 1666
