+++ test/test846.js
@@ -184,42 +184,6 @@ describe('Test ' + test + ' - non-numeric values for SUM, MIN and MAX', function
 		assert.deepEqual(res, [{a: 2}]);
 	});
 
-	it('SUM with Round function', function () {
-		var data = [
-			{a: null, b: 9, c: true, c2: 1, d: null, e: 'XYZ1', f: new Number(2)},
-			{
-				a: null,
-				b: 1,
-				c: false,
-				c2: false,
-				d: 5,
-				e: 'XYZ2',
-				f: new Number(11.25),
-			},
-		];
-		res = alasql(
-			`SELECT
-				SUM(a) AS a,
-				sum(b) as b,
-				sUm(c) as c,
-				sUm(c2) as c2,
-				SuM(d) as d,
-				SUM(e) as e,
-				SUM(ROUND(f)) as f
-			FROM ?`,
-			[data]
-		);
-		assert.deepEqual(res, [{a: null, b: 10, c: null, c2: 1, d: 5, e: null, f: 13}]);
-
-		var data = [[{a: null}]];
-		res = alasql(`SELECT SUM(a) AS a FROM ?`, data);
-		assert.deepEqual(res, [{a: null}]);
-
-		var data = [[{a: 2}]];
-		res = alasql(`SELECT SUM(a) AS a FROM ?`, data);
-		assert.deepEqual(res, [{a: 2}]);
-	});
-
 	it('SUM zero is zero', function () {
 		var data = [{v: 0}];
 		res = alasql(`select sum(v) as v from ?`, [data]);
